+++ plugin/api/constants.go
@@ -28,4 +28,4 @@ package api
 // API_VERSION is the version of the plugin API.
 //
 // This MUST be provided in the HandshakeResponse.
-const APIVersion int32 = 3
+const APIVersion int32 = 4