+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- No changes yet.
+
+
 ## [1.16.0] - 2019-01-22
 ### Added
 - Expose Thrift file names, package prefix, and Thrift root directory to
@@ -262,6 +266,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[Unreleased]: https://github.com/thriftrw/thriftrw-go/compare/v.16.0...HEAD
 [1.16.0]: https://github.com/thriftrw/thriftrw-go/compare/v.15.0...v1.16.0
 [1.15.0]: https://github.com/thriftrw/thriftrw-go/compare/v.14.0...v1.15.0
 [1.14.0]: https://github.com/thriftrw/thriftrw-go/compare/v.13.1...v1.14.0