+++ CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
-
+### Fixed
+- Bump API Version for ThriftRW plugins because the previous release contained
+  a significant change to the ThriftRW Plugin API.
 
 ## [1.16.0] - 2019-01-22
 ### Added