+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- No changes yet.
+
 ## [1.16.1] - 2019-01-23
 ### Fixed
 - Bump API Version for ThriftRW plugins because the previous release contained
@@ -267,6 +270,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[Unreleased]: https://github.com/thriftrw/thriftrw-go/compare/v1.16.1...HEAD
 [1.16.1]: https://github.com/thriftrw/thriftrw-go/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.15.0...v1.16.0
 [1.15.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.14.0...v1.15.0