+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.19.0] - 2019-04-26
 - Sets now support a `(go.type = "slice")` annotation to be generated as
   slices rather than maps.
 
@@ -279,7 +279,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/thriftrw/thriftrw-go/compare/v1.17.0...HEAD
+[1.19.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.18.0...v1.19.0
+[1.18.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.17.0...v1.18.0
 [1.17.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.16.1...v1.17.0
 [1.16.1]: https://github.com/thriftrw/thriftrw-go/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/thriftrw/thriftrw-go/compare/v1.15.0...v1.16.0